Struct google_pubsub1_beta2::ProjectMethods

pub struct ProjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Pubsub hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_pubsub1_beta2 as pubsub1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use pubsub1_beta2::Pubsub;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Pubsub::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `subscriptions_acknowledge(...)`, `subscriptions_create(...)`, `subscriptions_delete(...)`, `subscriptions_get(...)`, `subscriptions_list(...)`, `subscriptions_modify_ack_deadline(...)`, `subscriptions_modify_push_config(...)`, `subscriptions_pull(...)`, `topics_create(...)`, `topics_delete(...)`, `topics_get(...)`, `topics_list(...)`, `topics_publish(...)` and `topics_subscriptions_list(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

fn topics_list(&self, project: &str) -> ProjectTopicListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists matching topics.

Arguments

project - No description provided.

fn subscriptions_pull(&self, request: PullRequest, subscription: &str) -> ProjectSubscriptionPullCall<'a, C, A>

Create a builder to help you perform the following task:

Pulls messages from the server. Returns an empty list if there are no messages available in the backlog. The server may return UNAVAILABLE if there are too many concurrent pull requests pending for the given subscription.

Arguments

request - No description provided.

subscription - No description provided.

fn topics_create(&self, request: Topic, name: &str) -> ProjectTopicCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates the given topic with the given name.

Arguments

request - No description provided.

name - No description provided.

fn topics_delete(&self, topic: &str) -> ProjectTopicDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the topic with the given name. Returns NOT_FOUND if the topic does not exist. After a topic is deleted, a new topic may be created with the same name; this is an entirely new topic with none of the old configuration or subscriptions. Existing subscriptions to this topic are not deleted.

Arguments

topic - No description provided.

fn subscriptions_modify_push_config(&self, request: ModifyPushConfigRequest, subscription: &str) -> ProjectSubscriptionModifyPushConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies the PushConfig for a specified subscription. This may be used to change a push subscription to a pull one (signified by an empty PushConfig) or vice versa, or change the endpoint URL and other attributes of a push subscription. Messages will accumulate for delivery continuously through the call regardless of changes to the PushConfig.

Arguments

request - No description provided.

subscription - No description provided.

fn topics_publish(&self, request: PublishRequest, topic: &str) -> ProjectTopicPublishCall<'a, C, A>

Create a builder to help you perform the following task:

Adds one or more messages to the topic. Returns NOT_FOUND if the topic does not exist.

Arguments

request - No description provided.

topic - No description provided.

fn topics_get(&self, topic: &str) -> ProjectTopicGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the configuration of a topic.

Arguments

topic - No description provided.

fn subscriptions_modify_ack_deadline(&self, request: ModifyAckDeadlineRequest, subscription: &str) -> ProjectSubscriptionModifyAckDeadlineCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies the ack deadline for a specific message. This method is useful to indicate that more time is needed to process a message by the subscriber, or to make the message available for redelivery if the processing was interrupted.

Arguments

request - No description provided.

subscription - No description provided.

fn subscriptions_acknowledge(&self, request: AcknowledgeRequest, subscription: &str) -> ProjectSubscriptionAcknowledgeCall<'a, C, A>

Create a builder to help you perform the following task:

Acknowledges the messages associated with the ack tokens in the AcknowledgeRequest. The Pub/Sub system can remove the relevant messages from the subscription. Acknowledging a message whose ack deadline has expired may succeed, but such a message may be redelivered later. Acknowledging a message more than once will not result in an error.

Arguments

request - No description provided.

subscription - No description provided.

fn subscriptions_delete(&self, subscription: &str) -> ProjectSubscriptionDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an existing subscription. All pending messages in the subscription are immediately dropped. Calls to Pull after deletion will return NOT_FOUND. After a subscription is deleted, a new one may be created with the same name, but the new one has no association with the old subscription, or its topic unless the same topic is specified.

Arguments

subscription - No description provided.

fn subscriptions_create(&self, request: Subscription, name: &str) -> ProjectSubscriptionCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a subscription to a given topic for a given subscriber. If the subscription already exists, returns ALREADY_EXISTS. If the corresponding topic doesn't exist, returns NOT_FOUND. If the name is not provided in the request, the server will assign a random name for this subscription on the same project as the topic.

Arguments

request - No description provided.

name - No description provided.

fn subscriptions_get(&self, subscription: &str) -> ProjectSubscriptionGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the configuration details of a subscription.

Arguments

subscription - No description provided.

fn topics_subscriptions_list(&self, topic: &str) -> ProjectTopicSubscriptionListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the name of the subscriptions for this topic.

Arguments

topic - No description provided.

fn subscriptions_list(&self, project: &str) -> ProjectSubscriptionListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists matching subscriptions.

Arguments

project - No description provided.