Struct google-pubsub1_beta2::ProjectMethodsBuilder

pub struct ProjectMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Pubsub hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-pubsub1_beta2" as pubsub1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use pubsub1_beta2::Pubsub;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Pubsub::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `subscriptions_acknowledge(...)`, `subscriptions_create(...)`, `subscriptions_delete(...)`, `subscriptions_get(...)`, `subscriptions_list(...)`, `subscriptions_modify_ack_deadline(...)`, `subscriptions_modify_push_config(...)`, `subscriptions_pull(...)`, `topics_create(...)`, `topics_delete(...)`, `topics_get(...)`, `topics_list(...)`, `topics_publish(...)` and `topics_subscriptions_list(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, NC, A> ProjectMethodsBuilder<'a, C, NC, A>

fn topics_list(&self, project: &str) -> ProjectTopicListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists matching topics.

fn subscriptions_pull(&self, request: &PullRequest, subscription: &str) -> ProjectSubscriptionPullCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Pulls messages from the server. Returns an empty list if there are no messages available in the backlog. The server may return UNAVAILABLE if there are too many concurrent pull requests pending for the given subscription.

fn topics_create(&self, request: &Topic, name: &str) -> ProjectTopicCreateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates the given topic with the given name.

fn topics_delete(&self, topic: &str) -> ProjectTopicDeleteCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes the topic with the given name. All subscriptions to this topic are detached from the topic. Returns NOT_FOUND if the topic does not exist. After a topic is deleted, a new topic may be created with the same name; this is an entirely new topic with none of the old configuration or subscriptions.

fn subscriptions_modify_push_config(&self, request: &ModifyPushConfigRequest, subscription: &str) -> ProjectSubscriptionModifyPushConfigCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Modifies the PushConfig for a specified subscription. This may be used to change a push subscription to a pull one (signified by an empty PushConfig) or vice versa, or change the endpoint URL and other attributes of a push subscription. Messages will accumulate for delivery continuously through the call regardless of changes to the PushConfig.

fn topics_publish(&self, request: &PublishRequest, topic: &str) -> ProjectTopicPublishCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Adds one or more messages to the topic. Returns NOT_FOUND if the topic does not exist.

fn topics_get(&self, topic: &str) -> ProjectTopicGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the configuration of a topic.

fn subscriptions_modify_ack_deadline(&self, request: &ModifyAckDeadlineRequest, subscription: &str) -> ProjectSubscriptionModifyAckDeadlineCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Modifies the ack deadline for a specific message. This method is useful to indicate that more time is needed to process a message by the subscriber, or to make the message available for redelivery if the processing was interrupted.

fn subscriptions_acknowledge(&self, request: &AcknowledgeRequest, subscription: &str) -> ProjectSubscriptionAcknowledgeCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Acknowledges the messages associated with the ack tokens in the AcknowledgeRequest. The Pub/Sub system can remove the relevant messages from the subscription. Acknowledging a message whose ack deadline has expired may succeed, but such a message may be redelivered later. Acknowledging a message more than once will not result in an error.

fn subscriptions_delete(&self, subscription: &str) -> ProjectSubscriptionDeleteCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes an existing subscription. All pending messages in the subscription are immediately dropped. Calls to Pull after deletion will return NOT_FOUND. After a subscription is deleted, a new one may be created with the same name, but the new one has no association with the old subscription, or its topic unless the same topic is specified.

fn subscriptions_create(&self, request: &Subscription, name: &str) -> ProjectSubscriptionCreateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a subscription to a given topic for a given subscriber. If the subscription already exists, returns ALREADY_EXISTS. If the corresponding topic doesn't exist, returns NOT_FOUND. If the name is not provided in the request, the server will assign a random name for this subscription on the same project as the topic.

fn subscriptions_get(&self, subscription: &str) -> ProjectSubscriptionGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the configuration details of a subscription.

fn topics_subscriptions_list(&self, topic: &str) -> ProjectTopicSubscriptionListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists the name of the subscriptions for this topic.

fn subscriptions_list(&self, project: &str) -> ProjectSubscriptionListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists matching subscriptions.